+++ corpus/.clj-kondo/hooks/new.clj
@@ -5,4 +5,3 @@
   (api/reg-finding! (assoc (meta node)
                            :message (str "Interop is no good! " (api/generated-node? (first (:children node))))
                            :type :interop)))
-